requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astroid==3.1.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
isort==5.13.2
Markdown==3.5.2
mccabe==0.7.0
pillow==10.2.0
platformdirs==4.2.0
pydantic==2.6.2
pydantic_core==2.16.3
pylint==3.1.0
pytz==2024.1
setuptools==68.2.2
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.36.3
tomlkit==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.27.1
wheel==0.41.2
